Move object_json and update dependencies
